+++ app/src/organisms/ChooseRobotSlideout/index.tsx
@@ -24,7 +24,6 @@ import {
   SIZE_4,
   TEXT_ALIGN_CENTER,
 } from '@opentrons/components'
-import { ApiHostProvider } from '@opentrons/react-api-client'
 
 import {
   getConnectableRobots,
@@ -55,10 +54,8 @@ export function ChooseRobotSlideout(
 ): JSX.Element | null {
   const { t } = useTranslation(['protocol_details', 'shared'])
   const { storedProtocolData, showSlideout, onCloseClick, ...restProps } = props
-  const [createRunError, setCreateRunError] = React.useState<string | null>(
-    null
-  )
   const dispatch = useDispatch<Dispatch>()
+  const history = useHistory()
   const isScanning = useSelector((state: State) => getScanning(state))
 
   const unhealthyReachableRobots = useSelector((state: State) =>
@@ -73,6 +70,26 @@ export function ChooseRobotSlideout(
   const [selectedRobot, setSelectedRobot] = React.useState<Robot | null>(
     healthyReachableRobots[0] ?? null
   )
+  const {
+    createRunFromProtocolSource,
+    runCreationError,
+    reset: resetCreateRun,
+    isCreatingRun,
+  } = useCreateRunFromProtocol(
+    {
+      onSuccess: ({ data: runData }) => {
+        if (selectedRobot != null) {
+          history.push(
+            `/devices/${selectedRobot.name}/protocol-runs/${runData.id}`
+          )
+        }
+      },
+    },
+    selectedRobot != null ? { hostname: selectedRobot.ip } : null
+  )
+  const handleProceed: React.MouseEventHandler<HTMLButtonElement> = () => {
+    createRunFromProtocolSource({ files: srcFileObjects, protocolKey })
+  }
 
   const isSelectedRobotOnWrongVersionOfSoftware = [
     'upgrade',
@@ -121,19 +138,21 @@ export function ChooseRobotSlideout(
         protocol_name: protocolDisplayName,
       })}
       footer={
-        <ApiHostProvider
-          hostname={selectedRobot != null ? selectedRobot.ip : null}
+        <PrimaryButton
+          onClick={handleProceed}
+          width="100%"
+          disabled={
+            isCreatingRun ||
+            selectedRobot == null ||
+            isSelectedRobotOnWrongVersionOfSoftware
+          }
         >
-          <CreateRunButton
-            disabled={
-              selectedRobot == null || isSelectedRobotOnWrongVersionOfSoftware
-            }
-            protocolKey={protocolKey}
-            srcFileObjects={srcFileObjects}
-            setError={setCreateRunError}
-            robotName={selectedRobot != null ? selectedRobot.name : ''}
-          />
-        </ApiHostProvider>
+          {isCreatingRun ? (
+            <Icon name="ot-spinner" spin size={SIZE_1} />
+          ) : (
+            t('shared:proceed_to_setup')
+          )}
+        </PrimaryButton>
       }
       {...restProps}
     >
@@ -182,24 +201,27 @@ export function ChooseRobotSlideout(
                   robotModel="OT-2"
                   local={robot.local}
                   onClick={() => {
-                    setCreateRunError(null)
-                    setSelectedRobot(isSelected ? null : robot)
+                    if (!isCreatingRun) {
+                      resetCreateRun()
+                      setSelectedRobot(isSelected ? null : robot)
+                    }
                   }}
-                  isError={createRunError != null}
+                  isError={runCreationError != null}
                   isSelected={isSelected}
                   isOnDifferentSoftwareVersion={
                     isSelectedRobotOnWrongVersionOfSoftware
                   }
                 />
-                {createRunError != null && isSelected && (
+                {runCreationError != null && isSelected && (
                   <StyledText
                     as="label"
                     color={COLORS.errorText}
+                    css={{ 'overflow-wrap': 'anywhere' }}
                     display={DISPLAY_INLINE_BLOCK}
                     marginTop={`-${SPACING.spacing2}`}
                     marginBottom={SPACING.spacing3}
                   >
-                    {createRunError}
+                    {runCreationError}
                   </StyledText>
                 )}
               </Flex>
@@ -229,55 +251,5 @@ export function ChooseRobotSlideout(
         ) : null}
       </Flex>
     </Slideout>
-  )
-}
-
-interface CreateRunButtonProps
-  extends React.ComponentProps<typeof PrimaryButton> {
-  srcFileObjects: File[]
-  protocolKey: string
-  robotName: string
-  setError: (error: string | null) => void
-}
-function CreateRunButton(props: CreateRunButtonProps): JSX.Element {
-  const { t } = useTranslation('protocol_details')
-  const history = useHistory()
-  const {
-    protocolKey,
-    srcFileObjects,
-    robotName,
-    setError,
-    disabled,
-    ...buttonProps
-  } = props
-  const {
-    createRunFromProtocolSource,
-    runCreationError,
-    isCreatingRun,
-  } = useCreateRunFromProtocol({
-    onSuccess: ({ data: runData }) => {
-      history.push(`/devices/${robotName}/protocol-runs/${runData.id}`)
-    },
-  })
-
-  React.useEffect(() => {
-    if (runCreationError != null) {
-      setError(runCreationError)
-    }
-  }, [runCreationError, setError])
-
-  const handleClick: React.MouseEventHandler<HTMLButtonElement> = () => {
-    createRunFromProtocolSource({ files: srcFileObjects, protocolKey })
-  }
-
-  return (
-    <PrimaryButton
-      onClick={handleClick}
-      width="100%"
-      disabled={isCreatingRun || disabled}
-      {...buttonProps}
-    >
-      {t('proceed_to_setup')}
-    </PrimaryButton>
   )
 }