+++ app/src/organisms/Devices/ProtocolRun/ProtocolRunHeader.tsx
@@ -26,27 +26,23 @@ import {
   Flex,
   Icon,
   IconName,
-  Tooltip,
   useHoverTooltip,
   useInterval,
   ALIGN_CENTER,
   DIRECTION_COLUMN,
   DISPLAY_FLEX,
-  DIRECTION_ROW,
   JUSTIFY_CENTER,
   JUSTIFY_SPACE_BETWEEN,
   SIZE_1,
-  SIZE_3,
-  SIZE_4,
-  SIZE_5,
-  TEXT_TRANSFORM_UPPERCASE,
   BORDERS,
   COLORS,
   SPACING,
   TYPOGRAPHY,
   useConditionalConfirm,
+  SPACING_AUTO,
 } from '@opentrons/components'
 
+import { getBuildrootUpdateDisplayInfo } from '../../../redux/buildroot'
 import { ProtocolAnalysisErrorBanner } from './ProtocolAnalysisErrorBanner'
 import { ProtocolAnalysisErrorModal } from './ProtocolAnalysisErrorModal'
 import { Banner } from '../../../atoms/Banner'
@@ -54,6 +50,7 @@ import { PrimaryButton, SecondaryButton } from '../../../atoms/buttons'
 import { useTrackEvent } from '../../../redux/analytics'
 import { getIsHeaterShakerAttached } from '../../../redux/config'
 import { StyledText } from '../../../atoms/text'
+import { Tooltip } from '../../../atoms/Tooltip'
 import {
   useCloseCurrentRun,
   useCurrentRunId,
@@ -82,6 +79,7 @@ import { formatTimestamp } from '../utils'
 import { EMPTY_TIMESTAMP } from '../constants'
 
 import type { Run } from '@opentrons/api-client'
+import type { State } from '../../../redux/types'
 import type { HeaterShakerModule } from '../../../redux/modules/types'
 
 const EQUIPMENT_POLL_MS = 5000
@@ -114,11 +112,7 @@ function RunTimer({
   const runTime =
     startedAt != null ? formatInterval(startedAt, endTime) : EMPTY_TIMESTAMP
 
-  return (
-    <StyledText css={TYPOGRAPHY.pRegular} color={COLORS.darkBlack}>
-      {runTime}
-    </StyledText>
-  )
+  return <StyledText css={TYPOGRAPHY.pRegular}>{runTime}</StyledText>
 }
 
 export function ProtocolRunHeader({
@@ -276,6 +270,11 @@ export function ProtocolRunHeader({
     trackProtocolRunEvent({ name: 'runAgain' })
   }
 
+  const isRobotOnWrongVersionOfSoftware = ['upgrade', 'downgrade'].includes(
+    useSelector((state: State) => {
+      return getBuildrootUpdateDisplayInfo(state, robotName)
+    })?.autoUpdateAction
+  )
   const isRunControlButtonDisabled =
     (isCurrentRun && !isSetupComplete) ||
     isMutationLoading ||
@@ -284,7 +283,8 @@ export function ProtocolRunHeader({
     runStatus === RUN_STATUS_FINISHING ||
     runStatus === RUN_STATUS_PAUSE_REQUESTED ||
     runStatus === RUN_STATUS_STOP_REQUESTED ||
-    runStatus === RUN_STATUS_BLOCKED_BY_OPEN_DOOR
+    runStatus === RUN_STATUS_BLOCKED_BY_OPEN_DOOR ||
+    isRobotOnWrongVersionOfSoftware
 
   let handleButtonClick = (): void => {}
   let buttonIconName: IconName | null = null
@@ -330,6 +330,8 @@ export function ProtocolRunHeader({
     disableReason = t('setup_incomplete')
   } else if (isRobotBusy) {
     disableReason = t('robot_is_busy')
+  } else if (isRobotOnWrongVersionOfSoftware) {
+    disableReason = t('shared:a_software_update_is_available')
   }
 
   const buttonIcon =
@@ -370,85 +372,90 @@ export function ProtocolRunHeader({
     closeCurrentRun()
   }
 
-  const ClearProtocolBanner = (): JSX.Element | null => {
-    switch (runStatus) {
-      case RUN_STATUS_FAILED: {
-        return (
-          <Banner type="error" onCloseClick={handleClearClick}>
-            <Flex justifyContent={JUSTIFY_SPACE_BETWEEN} width="100%">
-              {`${t('run_failed')}.`}
-            </Flex>
-          </Banner>
-        )
-      }
-      case RUN_STATUS_SUCCEEDED: {
-        return (
-          <Banner type="success" onCloseClick={handleClearClick}>
-            <Flex justifyContent={JUSTIFY_SPACE_BETWEEN} width="100%">
-              {`${t('run_completed')}.`}
-            </Flex>
-          </Banner>
-        )
-      }
+  let clearProtocolBanner: JSX.Element | null
+  switch (runStatus) {
+    case RUN_STATUS_FAILED: {
+      clearProtocolBanner = (
+        <Banner type="error" onCloseClick={handleClearClick}>
+          <Flex justifyContent={JUSTIFY_SPACE_BETWEEN} width="100%">
+            {`${t('run_failed')}.`}
+          </Flex>
+        </Banner>
+      )
+      break
     }
-    return null
+    case RUN_STATUS_SUCCEEDED: {
+      clearProtocolBanner = (
+        <Banner type="success" onCloseClick={handleClearClick}>
+          <Flex justifyContent={JUSTIFY_SPACE_BETWEEN} width="100%">
+            {`${t('run_completed')}.`}
+          </Flex>
+        </Banner>
+      )
+      break
+    }
+    default:
+      clearProtocolBanner = null
   }
 
-  const ProtocolRunningContent = (): JSX.Element | null =>
+  const protocolRunningContent: JSX.Element | null =
     runStatus != null ? (
-      <Flex
+      <Box
         backgroundColor={COLORS.lightGrey}
-        justifyContent={JUSTIFY_SPACE_BETWEEN}
+        display="grid"
+        gridTemplateColumns="4fr 6fr 4fr"
         padding={SPACING.spacing3}
       >
-        <Flex gridGap={SPACING.spacing6}>
-          <Box>
-            <StyledText
-              textTransform={TEXT_TRANSFORM_UPPERCASE}
-              color={COLORS.darkGreyEnabled}
-              css={TYPOGRAPHY.h6Default}
-              paddingBottom={SPACING.spacing2}
-            >
-              {t('protocol_start')}
-            </StyledText>
-            <StyledText
-              css={TYPOGRAPHY.pRegular}
-              color={COLORS.darkBlack}
-              id="ProtocolRunHeader_protocolStart"
-            >
-              {startedAtTimestamp}
-            </StyledText>
-          </Box>
-          <Box>
-            <StyledText
-              textTransform={TEXT_TRANSFORM_UPPERCASE}
-              color={COLORS.darkGreyEnabled}
-              css={TYPOGRAPHY.h6Default}
-              paddingBottom={SPACING.spacing2}
-            >
-              {t('protocol_end')}
-            </StyledText>
-            <StyledText
-              css={TYPOGRAPHY.pRegular}
-              color={COLORS.darkBlack}
-              id="ProtocolRunHeader_protocolEnd"
-            >
-              {completedAtTimestamp}
-            </StyledText>
-          </Box>
-        </Flex>
-        {showCancelButton ? (
-          <SecondaryButton
-            color={COLORS.errorText}
-            padding={`${SPACING.spacingSM} ${SPACING.spacing4}`}
-            onClick={handleCancelClick}
-            id="ProtocolRunHeader_cancelRunButton"
-            disabled={isClosingCurrentRun}
+        <Box>
+          <StyledText
+            textTransform={TYPOGRAPHY.textTransformUppercase}
+            color={COLORS.darkGreyEnabled}
+            css={TYPOGRAPHY.h6Default}
+            paddingBottom={SPACING.spacing2}
           >
-            {t('cancel_run')}
-          </SecondaryButton>
-        ) : null}
-      </Flex>
+            {t('protocol_start')}
+          </StyledText>
+          <StyledText
+            css={TYPOGRAPHY.pRegular}
+            id="ProtocolRunHeader_protocolStart"
+          >
+            {startedAtTimestamp}
+          </StyledText>
+        </Box>
+        <Box>
+          <StyledText
+            textTransform={TYPOGRAPHY.textTransformUppercase}
+            color={COLORS.darkGreyEnabled}
+            css={TYPOGRAPHY.h6Default}
+            paddingBottom={SPACING.spacing2}
+          >
+            {t('protocol_end')}
+          </StyledText>
+          <StyledText
+            css={TYPOGRAPHY.pRegular}
+            id="ProtocolRunHeader_protocolEnd"
+          >
+            {completedAtTimestamp}
+          </StyledText>
+        </Box>
+        <Box marginLeft={SPACING_AUTO}>
+          {showCancelButton != null && (
+            <SecondaryButton
+              style={{
+                color: `${COLORS.errorText}`,
+                backgroundColor: 'none',
+                borderColor: `${COLORS.error}`,
+              }}
+              padding={`${SPACING.spacingSM} ${SPACING.spacing4}`}
+              onClick={handleCancelClick}
+              id="ProtocolRunHeader_cancelRunButton"
+              disabled={isClosingCurrentRun}
+            >
+              {t('cancel_run')}
+            </SecondaryButton>
+          )}
+        </Box>
+      </Box>
     ) : null
 
   return (
@@ -501,7 +508,7 @@ export function ProtocolRunHeader({
           </StyledText>
         )}
       </Flex>
-      {analysisErrors && analysisErrors?.length > 0 && (
+      {analysisErrors != null && analysisErrors?.length > 0 && (
         <ProtocolAnalysisErrorBanner errors={analysisErrors} />
       )}
       {runStatus === RUN_STATUS_BLOCKED_BY_OPEN_DOOR ? (
@@ -510,29 +517,28 @@ export function ProtocolRunHeader({
       {runStatus === RUN_STATUS_STOPPED ? (
         <Banner type="warning">{`${t('run_canceled')}.`}</Banner>
       ) : null}
-      {isCurrentRun ? <ClearProtocolBanner /> : null}
-      <Flex justifyContent={JUSTIFY_SPACE_BETWEEN}>
-        <Box minWidth={SIZE_4}>
+      {isCurrentRun ? clearProtocolBanner : null}
+      <Box display="grid" gridTemplateColumns="4fr 3fr 3fr 4fr">
+        <Box>
           <StyledText
-            textTransform={TEXT_TRANSFORM_UPPERCASE}
+            textTransform={TYPOGRAPHY.textTransformUppercase}
             color={COLORS.darkGreyEnabled}
             css={TYPOGRAPHY.h6Default}
             paddingBottom={SPACING.spacing2}
           >
-            {t('run_id')}
+            {t('run')}
           </StyledText>
           {/* this is the createdAt timestamp, not the run id */}
           <StyledText
             css={TYPOGRAPHY.pRegular}
-            color={COLORS.darkBlack}
             id="ProtocolRunHeader_runRecordId"
           >
             {createdAtTimestamp}
           </StyledText>
         </Box>
-        <Box minWidth={SIZE_3}>
+        <Box>
           <StyledText
-            textTransform={TEXT_TRANSFORM_UPPERCASE}
+            textTransform={TYPOGRAPHY.textTransformUppercase}
             color={COLORS.darkGreyEnabled}
             css={TYPOGRAPHY.h6Default}
             paddingBottom={SPACING.spacing2}
@@ -560,16 +566,15 @@ export function ProtocolRunHeader({
             ) : null}
             <StyledText
               css={TYPOGRAPHY.pRegular}
-              color={COLORS.darkBlack}
               id="ProtocolRunHeader_runStatus"
             >
               {runStatus != null ? t(`status_${runStatus}`) : ''}
             </StyledText>
           </Flex>
         </Box>
-        <Box minWidth={SIZE_3}>
+        <Box>
           <StyledText
-            textTransform={TEXT_TRANSFORM_UPPERCASE}
+            textTransform={TYPOGRAPHY.textTransformUppercase}
             color={COLORS.darkGreyEnabled}
             css={TYPOGRAPHY.h6Default}
             paddingBottom={SPACING.spacing2}
@@ -590,12 +595,7 @@ export function ProtocolRunHeader({
             startRun={play}
           />
         )}
-        <Flex
-          justifyContent={'flex-end'}
-          flexDirection={DIRECTION_ROW}
-          gridGap={SPACING.spacingSM}
-          width={SIZE_5}
-        >
+        <Box marginLeft={SPACING_AUTO}>
           <PrimaryButton
             justifyContent={JUSTIFY_CENTER}
             alignItems={ALIGN_CENTER}
@@ -611,11 +611,11 @@ export function ProtocolRunHeader({
             <StyledText css={TYPOGRAPHY.pSemiBold}>{buttonText}</StyledText>
           </PrimaryButton>
           {disableReason != null && (
-            <Tooltip {...tooltipProps}>{disableReason}</Tooltip>
+            <Tooltip tooltipProps={tooltipProps}>{disableReason}</Tooltip>
           )}
-        </Flex>
-      </Flex>
-      <ProtocolRunningContent />
+        </Box>
+      </Box>
+      {protocolRunningContent}
       {showConfirmCancelModal ? (
         <ConfirmCancelModal
           onClose={() => setShowConfirmCancelModal(false)}