+++ app/src/organisms/ModuleCard/HeaterShakerSlideout.tsx
@@ -119,7 +119,10 @@ export const HeaterShakerSlideout = (
     !configHasHeaterShakerAttached
   )
 
-  const sendSetTemperatureOrShakeCommand = (): void => {
+  const sendSetTemperatureOrShakeCommand: React.MouseEventHandler<HTMLInputElement> = e => {
+    e.preventDefault()
+    e.stopPropagation()
+
     if (hsValue != null && !isSetShake) {
       const setTempCommand: HeaterShakerStartSetTargetTemperatureCreateCommand = {
         commandType: 'heaterShaker/setTargetTemperature',
@@ -149,8 +152,12 @@ export const HeaterShakerSlideout = (
         })
       }
     }
-    isSetShake ? confirmAttachment() : setHsValue(null)
-    onCloseClick()
+    if (isSetShake) {
+      confirmAttachment()
+    } else {
+      setHsValue(null)
+      onCloseClick()
+    }
   }
 
   let errorMessage