+++ components/src/ui-style-constants/typography.ts
@@ -121,13 +121,12 @@ export const linkPSemiBold = css`
   }
 `
 
-export const darkLinkPSemiBold = css`
-  font-size: ${fontSizeP};
+export const darkLinkLabelSemiBold = css`
+  font-size: ${fontSizeLabel};
   font-weight: ${fontWeightSemiBold};
   line-height: ${lineHeight20};
-  color: ${COLORS.darkBlack};
-
+  color: ${COLORS.darkGreyEnabled};
   &:hover {
-    opacity: 1;
+    color: ${COLORS.darkBlack};
   }
 `